+++ include/ffapi.h
@@ -39,7 +39,7 @@ ffapi_pix_fmt_filter ffapi_pixfmts_8bit_pel, ffapi_pixfmts_32_bit_float_pel;
 void       ffapi_parse_color_props(FFColorProperties* c, const char* props);
 FFContext* ffapi_open_input (const char* file, const char* options,
                              const char* format, FFColorProperties* color_props, ffapi_pix_fmt_filter*,
-                             unsigned long* components, unsigned long (*widths)[4], unsigned long (*heights)[4], unsigned long* frames,
+                             unsigned long* components, unsigned long (*widths)[4], unsigned long (*heights)[4], uint64_t* frames,
                              AVRational* rate, bool calc_frames);
 FFContext* ffapi_open_output(const char* file, const char* options,
                              const char* format, const char* encoder, enum AVCodecID preferred_encoder,
@@ -49,7 +49,7 @@ AVFrame*  ffapi_alloc_frame(FFContext*);
 void      ffapi_free_frame (AVFrame*);
 void      ffapi_clear_frame(AVFrame*);
 int       ffapi_read_frame (FFContext*, AVFrame*);
-size_t    ffapi_seek_frame (FFContext*, size_t offset, void (*progress)(size_t));
+uint64_t  ffapi_seek_frame (FFContext*, uint64_t offset, void (*progress)(uint64_t));
 int       ffapi_write_frame(FFContext*, AVFrame*);
 int       ffapi_close(FFContext*);
